+++ src/EmbeddedNativeLibrary.cs
@@ -1,6 +1,5 @@
 ﻿using System;
 using System.Collections.Generic;
-using System.ComponentModel;
 using System.Diagnostics;
 using System.IO;
 using System.Linq;
@@ -17,7 +16,7 @@ namespace RockLib.Interop
     /// embedded as a resource in the same assembly that <see cref="EmbeddedNativeLibrary"/>
     /// is defined.
     /// </summary>
-    internal sealed class EmbeddedNativeLibrary : IDisposable
+    internal sealed partial class EmbeddedNativeLibrary : IDisposable
     {
         private const bool _defaultPreferEmbeddedOverInstalled = true;
 
@@ -492,463 +491,5 @@ namespace RockLib.Interop
                 _libraryLoader.FreeLibrary(_libraryPointer.Value);
             }
         }
-
-        private interface ILibraryLoader
-        {
-            string[] CandidateWritableLocations { get; }
-            IEnumerable<string> GetInstallPathCandidates(string libraryName);
-            MaybeIntPtr LoadLibrary(string libraryPath);
-            void FreeLibrary(IntPtr libraryPointer);
-            MaybeIntPtr GetFunctionPointer(IntPtr libraryPointer, string functionName);
-        }
-
-        private class MaybeIntPtr
-        {
-            public MaybeIntPtr(IntPtr value)
-            {
-                Debug.Assert(value != IntPtr.Zero, "MaybeIntPtr.ctor(IntPtr value): value must be non-zero.");
-                Value = value;
-            }
-
-            public MaybeIntPtr(Exception[] exceptions)
-            {
-                Debug.Assert(exceptions != null, "MaybeIntPtr.ctor(Exception[] exceptions): exceptions parameter must not be null.");
-                Debug.Assert(exceptions.Length >= 1, "MaybeIntPtr.ctor(Exception[] exceptions): exceptions must contain at least one element.");
-                Exceptions = exceptions;
-            }
-
-            public IntPtr Value { get; private set; }
-            public Exception[] Exceptions { get; private set; }
-            public bool HasValue { get { return Value != IntPtr.Zero; } }
-        }
-
-        private class WindowsLibraryLoader : ILibraryLoader
-        {
-            private static readonly string[] _candidateWritableLocations;
-
-            static WindowsLibraryLoader()
-            {
-                var candidateLocations = new List<string>();
-                var localAppData = Environment.GetEnvironmentVariable("LocalAppData");
-                if (!string.IsNullOrEmpty(localAppData))
-                {
-                    candidateLocations.Add(localAppData);
-                }
-
-                var tmpDirectory = Environment.GetEnvironmentVariable("TMP");
-                if (!string.IsNullOrEmpty(tmpDirectory))
-                {
-                    candidateLocations.Add(tmpDirectory);
-                }
-
-                var tempDirectory = Environment.GetEnvironmentVariable("TEMP");
-                if (!string.IsNullOrEmpty(tempDirectory))
-                {
-                    candidateLocations.Add(tempDirectory);
-                }
-
-                _candidateWritableLocations = candidateLocations.ToArray();
-            }
-
-            public string[] CandidateWritableLocations { get { return _candidateWritableLocations; } }
-
-            public IEnumerable<string> GetInstallPathCandidates(string libraryName)
-            {
-                var fullName = libraryName + ".dll";
-
-                var assembly = Assembly.GetEntryAssembly() ?? Assembly.GetExecutingAssembly();
-                var potentialInstallPath = Path.Combine(Path.GetDirectoryName(assembly.Location), fullName);
-                if (File.Exists(potentialInstallPath))
-                {
-                    yield return potentialInstallPath;
-                }
-
-                yield return fullName;
-            }
-
-            public MaybeIntPtr LoadLibrary(string libraryPath)
-            {
-                var libraryPointer = NativeMethods.LoadLibrary(libraryPath);
-
-                if (libraryPointer != IntPtr.Zero)
-                {
-                    return new MaybeIntPtr(libraryPointer);
-                }
-
-                var exceptions = new List<Exception>();
-
-                exceptions.Add(new Win32Exception());
-
-                libraryPointer = NativeMethods.LoadLibraryEx(libraryPath, IntPtr.Zero, LoadLibraryFlags.LOAD_LIBRARY_SEARCH_DLL_LOAD_DIR | LoadLibraryFlags.LOAD_LIBRARY_SEARCH_DEFAULT_DIRS);
-
-                if (libraryPointer != IntPtr.Zero)
-                {
-                    return new MaybeIntPtr(libraryPointer);
-                }
-
-                exceptions.Add(new Win32Exception());
-
-                var originalPathVariable = Environment.GetEnvironmentVariable("PATH");
-                var pathVariable = originalPathVariable + ";" + Path.GetDirectoryName(libraryPath);
-                Environment.SetEnvironmentVariable("PATH", pathVariable);
-
-                libraryPointer = NativeMethods.LoadLibrary(libraryPath);
-
-                Environment.SetEnvironmentVariable("PATH", originalPathVariable);
-
-                if (libraryPointer != IntPtr.Zero)
-                {
-                    return new MaybeIntPtr(libraryPointer);
-                }
-
-                exceptions.Add(new Win32Exception());
-
-                return new MaybeIntPtr(exceptions.ToArray());
-            }
-
-            public void FreeLibrary(IntPtr libraryPointer)
-            {
-                NativeMethods.FreeLibrary(libraryPointer);
-            }
-
-            public MaybeIntPtr GetFunctionPointer(IntPtr libraryPointer, string functionName)
-            {
-                var functionPointer = NativeMethods.GetProcAddress(libraryPointer, functionName);
-
-                if (functionPointer != IntPtr.Zero)
-                {
-                    return new MaybeIntPtr(functionPointer);
-                }
-
-                return new MaybeIntPtr(new Exception[] { new Win32Exception() });
-            }
-
-            private static class NativeMethods
-            {
-                [DllImport("kernel32.dll", EntryPoint = "LoadLibraryEx", BestFitMapping = false, ThrowOnUnmappableChar = true, SetLastError = true)]
-                public static extern IntPtr LoadLibraryEx([MarshalAs(UnmanagedType.LPStr)] string lpFileName, IntPtr hReservedNull, LoadLibraryFlags dwFlags);
-
-                [DllImport("kernel32.dll", EntryPoint = "LoadLibrary", BestFitMapping = false, ThrowOnUnmappableChar = true, SetLastError = true)]
-                public static extern IntPtr LoadLibrary([MarshalAs(UnmanagedType.LPStr)] string lpLibFileName);
-
-                [DllImport("kernel32.dll", EntryPoint = "GetProcAddress", BestFitMapping = false, ThrowOnUnmappableChar = true, SetLastError = true)]
-                public static extern IntPtr GetProcAddress(IntPtr hModule, [MarshalAs(UnmanagedType.LPStr)] string lpProcName);
-
-                [DllImport("kernel32.dll", EntryPoint = "FreeLibrary", SetLastError = true)]
-                public static extern bool FreeLibrary(IntPtr hModule);
-            }
-
-            [Flags]
-            private enum LoadLibraryFlags : uint
-            {
-                DONT_RESOLVE_DLL_REFERENCES = 0x00000001,
-                LOAD_LIBRARY_AS_DATAFILE = 0x00000002,
-                LOAD_WITH_ALTERED_SEARCH_PATH = 0x00000008,
-                LOAD_IGNORE_CODE_AUTHZ_LEVEL = 0x00000010,
-                LOAD_LIBRARY_AS_IMAGE_RESOURCE = 0x00000020,
-                LOAD_LIBRARY_AS_DATAFILE_EXCLUSIVE = 0x00000040,
-                LOAD_LIBRARY_SEARCH_DLL_LOAD_DIR = 0x00000100,
-                LOAD_LIBRARY_SEARCH_APPLICATION_DIR = 0x00000200,
-                LOAD_LIBRARY_SEARCH_USER_DIRS = 0x00000400,
-                LOAD_LIBRARY_SEARCH_SYSTEM32 = 0x00000800,
-                LOAD_LIBRARY_SEARCH_DEFAULT_DIRS = 0x00001000,
-            }
-        }
-
-        private class UnixLibraryLoader : ILibraryLoader
-        {
-            private static readonly string[] _candidateWritableLocations = new[] { "/tmp", "/var/tmp" };
-
-            private readonly bool _isMac;
-
-            public UnixLibraryLoader(bool isMac)
-            {
-                _isMac = isMac;
-            }
-
-            public string[] CandidateWritableLocations { get { return _candidateWritableLocations; } }
-
-            public IEnumerable<string> GetInstallPathCandidates(string libraryName)
-            {
-                var fullName = libraryName + (_isMac ? ".dylib" : ".so");
-
-                var assembly = Assembly.GetEntryAssembly() ?? Assembly.GetExecutingAssembly();
-                var potentialInstallPath = Path.Combine(Path.GetDirectoryName(assembly.Location), fullName);
-                if (File.Exists(potentialInstallPath))
-                {
-                    yield return potentialInstallPath;
-                }
-
-                yield return fullName;
-            }
-
-            public MaybeIntPtr LoadLibrary(string libraryPath)
-            {
-                var libraryPointer = dlopen(libraryPath, dlopenFlags.RTLD_LAZY | dlopenFlags.RTLD_GLOBAL);
-
-                if (libraryPointer != IntPtr.Zero)
-                {
-                    return new MaybeIntPtr(libraryPointer);
-                }
-
-                var error = dlerror();
-                if (string.IsNullOrEmpty(error))
-                {
-                    error = "Null pointer was returned from dlopen.";
-                }
-                return new MaybeIntPtr(new[] { new Exception(error) });
-            }
-
-            public void FreeLibrary(IntPtr libraryPointer)
-            {
-                dlclose(libraryPointer);
-            }
-
-            public MaybeIntPtr GetFunctionPointer(IntPtr libraryPointer, string functionName)
-            {   
-                var functionPointer = dlsym(libraryPointer, functionName);
-
-                if (functionPointer != IntPtr.Zero)
-                {
-                    return new MaybeIntPtr(functionPointer);
-                }
-
-                return new MaybeIntPtr(new Exception[] { new Exception(dlerror()) });
-            }
-
-            private IntPtr dlopen(string filename, dlopenFlags flags) =>
-                 _isMac ? Mac.NativeMethods.dlopen(filename, flags) : Linux.NativeMethods.dlopen(filename, flags);
-
-            private string dlerror() =>
-                _isMac ? Mac.NativeMethods.dlerror() : Linux.NativeMethods.dlerror();
-
-            private IntPtr dlsym(IntPtr handle, string symbol) =>
-                _isMac ? Mac.NativeMethods.dlsym(handle, symbol) : Linux.NativeMethods.dlsym(handle, symbol);
-
-            private IntPtr dlclose(IntPtr handle) =>
-                _isMac ? Mac.NativeMethods.dlclose(handle) : Linux.NativeMethods.dlclose(handle);
-
-            private static class Mac
-            {
-                internal static class NativeMethods
-                {
-                    [DllImport("libSystem.dylib")]
-                    public static extern IntPtr dlopen(string filename, dlopenFlags flags);
-
-                    [DllImport("libSystem.dylib")]
-                    public static extern string dlerror();
-
-                    [DllImport("libSystem.dylib")]
-                    public static extern IntPtr dlsym(IntPtr handle, string symbol);
-
-                    [DllImport("libSystem.dylib")]
-                    public static extern IntPtr dlclose(IntPtr handle);
-                }
-            }
-
-            private static class Linux
-            {
-                internal static class NativeMethods
-                {// libdl.so libcoreclr.so
-                    [DllImport("libdl.so")]
-                    public static extern IntPtr dlopen(string filename, dlopenFlags flag);
-
-                    [DllImport("libdl.so")]
-                    public static extern string dlerror();
-
-                    [DllImport("libdl.so")]
-                    public static extern IntPtr dlsym(IntPtr handle, string name);
-
-                    [DllImport("libdl.so")]
-                    public static extern IntPtr dlclose(IntPtr handle);
-                }
-            }
-            
-            [Flags]
-            private enum dlopenFlags
-            {
-                RTLD_LAZY = 0x1,
-                RTLD_NOW = 0x2,
-                RTLD_LOCAL = 0x4,
-                RTLD_GLOBAL = 0x8,
-                RTLD_NOLOAD = 0x10,
-                RTLD_NODELETE = 0x80,
-            }
-        }
-
-        private class NullLibraryLoader : ILibraryLoader
-        {
-            private static readonly string[] _empty = new string[0];
-
-            public string[] CandidateWritableLocations
-            {
-                get { return _empty; }
-            }
-
-            public IEnumerable<string> GetInstallPathCandidates(string libraryName) { return _empty; }
-
-            public void FreeLibrary(IntPtr libraryPointer)
-            {
-            }
-
-            public MaybeIntPtr GetFunctionPointer(IntPtr libraryPointer, string functionName)
-            {
-                throw new NotImplementedException();
-            }
-
-            public MaybeIntPtr LoadLibrary(string libraryPath)
-            {
-                return new MaybeIntPtr(IntPtr.Zero);
-            }
-        }
-
-        private enum RuntimeOS
-        {
-            Windows,
-            Mac,
-            Linux
-        }
-    }
-
-    /// <summary>
-    /// Contains resource names for the DLLs that are embedded in this assembly. The DLLs
-    /// must all be of the same architecture (x86 or x64).
-    /// </summary>
-    internal sealed class DllInfo
-    {
-        private static readonly IReadOnlyCollection<string> _assemblyManifestResourceNames = typeof(DllInfo).GetTypeInfo().Assembly.GetManifestResourceNames().Select(n => n.ToLowerInvariant()).ToList().AsReadOnly();
-
-        private readonly TargetRuntime _targetRuntime;
-        private readonly string _resourceName;
-        private readonly string[] _additionalResourceNames;
-
-        /// <summary>
-        /// Initializes a new instance of the <see cref="DllInfo"/> class, assuming the target runtime to be
-        /// <see cref="TargetRuntime.Windows"/>.
-        /// </summary>
-        /// <param name="resourceName">The resource name of the main DLL to be loaded.</param>
-        /// <param name="additionalResourceNames">
-        /// The resource names of any additional DLLs that neede to be loaded.
-        /// </param>
-        /// <exception cref="ArgumentNullException">
-        /// <paramref name="resourceName"/> is null.
-        /// </exception>
-        /// <exception cref="ArgumentException">
-        /// <paramref name="resourceName"/> is empty.
-        /// or
-        /// <paramref name="resourceName"/> is not found in this assembly's manifest resource names.
-        /// or
-        /// <paramref name="additionalResourceNames"/> has any null elements.
-        /// or
-        /// <paramref name="additionalResourceNames"/> has any empty elements.
-        /// or
-        /// <paramref name="additionalResourceNames"/> has any elements that are not found in this assembly's manifest resource names.
-        /// </exception>
-        public DllInfo(string resourceName, params string[] additionalResourceNames)
-            : this(TargetRuntime.Windows, resourceName, additionalResourceNames)
-        {
-        }
-
-        /// <summary>
-        /// Initializes a new instance of the <see cref="DllInfo"/> class.
-        /// </summary>
-        /// <param name="targetRuntime">The runtime that this <see cref="DllInfo"/> targets.</param>
-        /// <param name="resourceName">The resource name of the main DLL to be loaded.</param>
-        /// <param name="additionalResourceNames">
-        /// The resource names of any additional DLLs that neede to be loaded.
-        /// </param>
-        /// <exception cref="ArgumentNullException">
-        /// <paramref name="resourceName"/> is null.
-        /// </exception>
-        /// <exception cref="ArgumentException">
-        /// <paramref name="resourceName"/> is empty.
-        /// or
-        /// <paramref name="resourceName"/> is not found in this assembly's manifest resource names.
-        /// or
-        /// <paramref name="additionalResourceNames"/> has any null elements.
-        /// or
-        /// <paramref name="additionalResourceNames"/> has any empty elements.
-        /// or
-        /// <paramref name="additionalResourceNames"/> has any elements that are not found in this assembly's manifest resource names.
-        /// </exception>
-        public DllInfo(TargetRuntime targetRuntime, string resourceName, params string[] additionalResourceNames)
-        {
-            if (resourceName == null) throw new ArgumentNullException("resourceName");
-            if (resourceName == "") throw new ArgumentException("'resourceName' must not be empty.", "resourceName");
-            if (!_assemblyManifestResourceNames.Contains(resourceName.ToLowerInvariant()))
-                throw new ArgumentException(string.Format("Resource '{0}' was not found in the assembly manifest resource names: {1}",
-                    resourceName, string.Join(", ", _assemblyManifestResourceNames.Select(n => "'" + n + "'"))), "resourceName");
-
-            if (additionalResourceNames != null)
-            {
-                foreach (var additionalResourceName in additionalResourceNames)
-                {
-                    if (additionalResourceName == null) throw new ArgumentException("Elements of 'additionalResourceNames' must not be null.", "additionalResourceNames");
-                    if (additionalResourceName == "") throw new ArgumentException("Elements of 'additionalResourceNames' must not be empty.", "additionalResourceNames");
-                    if (!_assemblyManifestResourceNames.Contains(additionalResourceName.ToLowerInvariant()))
-                        throw new ArgumentException(string.Format("Additional resource '{0}' was not found in the assembly manifest resource names: {1}",
-                            additionalResourceName, string.Join(", ", _assemblyManifestResourceNames.Select(n => "'" + n + "'"))), "additionalResourceNames");
-                }
-            }
-
-            _targetRuntime = targetRuntime;
-            _resourceName = resourceName;
-            _additionalResourceNames = additionalResourceNames ?? new string[0];
-        }
-
-        /// <summary>
-        /// Gets the runtime that this <see cref="DllInfo"/> targets.
-        /// </summary>
-        public TargetRuntime TargetRuntime
-        {
-            get { return _targetRuntime; }
-        }
-
-        /// <summary>
-        /// Gets the resource name of the main DLL to be loaded.
-        /// </summary>
-        public string ResourceName
-        {
-            get { return _resourceName; }
-        }
-
-        /// <summary>
-        /// Gets the resource names of any additional DLLs that need to be loaded.
-        /// </summary>
-        public string[] AdditionalResourceNames
-        {
-            get { return _additionalResourceNames; }
-        }
-    }
-
-    /// <summary>
-    /// Defines the target runtimes that are supported by the <see cref="DllInfo"/> class.
-    /// </summary>
-    internal enum TargetRuntime
-    {
-        /// <summary>
-        /// A windows environment. Whether it is 32-bit or 64-bit is unspecified.
-        /// </summary>
-        Windows,
-
-        /// <summary>
-        /// A windows 32-bit environment.
-        /// </summary>
-        Win32,
-
-        /// <summary>
-        /// A Windows 64-bit environment.
-        /// </summary>
-        Win64,
-
-        /// <summary>
-        /// A Mac environment.
-        /// </summary>
-        Mac,
-
-        /// <summary>
-        /// A Linux environment.
-        /// </summary>
-        Linux,
     }
 }
