+++ src/RuntimeOS.cs
@@ -1,0 +1,12 @@
+﻿namespace RockLib.Interop
+{
+    partial class EmbeddedNativeLibrary
+    {
+        private enum RuntimeOS
+        {
+            Windows,
+            Mac,
+            Linux
+        }
+    }
+}
